cast type of INSIDE_AREA to match segIndex

+++ src/operation/distance/GeometryLocation.cpp
@@ -46,7 +46,7 @@ GeometryLocation::GeometryLocation(const Geometry* newComponent, const Coordinat
 {
     component = newComponent;
     inside_area = true;
-    segIndex = INSIDE_AREA;
+    segIndex = (size_t) INSIDE_AREA;
     pt = newPt;
 }
